New branch gl-master

+++ USER/APP/key.h
@@ -19,6 +19,6 @@
 void Delay_ms(uint16_t ms);
 int8_t KeyPress(GPIO_TypeDef* GPIOx,uint16_t kpin);
 int8_t KeyDown(GPIO_TypeDef* GPIOx,uint16_t kpin);
-int16_t KeyDown_Ex(GPIO_TypeDef* GPIOx,uint16_t kpin);
+int8_t KeyDown_Ex(GPIO_TypeDef* GPIOx,uint16_t kpin);
 #endif
